updating Java web dev projects folder after resolving errors
