--- classes-part-2/studio/src/main/java/org/launchcode/MenuItem.java
@@ -1,32 +1,0 @@
-package org.launchcode;
-
-public class MenuItem {
-    private double price;
-    private String description;
-    private String category;
-    private boolean isNew;
-
-    public MenuItem(double p, String d, String c, boolean iN) {
-        this.price = p;
-        this.description = d;
-        this.category = c;
-        this.isNew = iN;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public void setCategory(String category) {
-        this.category = category;
-    }
-
-    public void setNew(boolean aNew) {
-        isNew = aNew;
-    }
-}
-
