+++ classes-part-one/studio/restaurant-menu/src/main/java/org/launchcode/Menu.java
@@ -1,0 +1,16 @@
+package org.launchcode;
+
+import java.time.LocalDate;
+import java.util.ArrayList;
+
+public class Menu {
+
+    private ArrayList<MenuItem> menuItems = new ArrayList<>();
+    private LocalDate lastUpdated;
+
+//Default Constructor
+public ArrayList<MenuItem> getMenuItems() {
+    return menuItems;
+}
+
+}
