+++ classes-part-one/studio/restaurant-menu/src/main/java/org/launchcode/MenuItem.java
@@ -1,0 +1,49 @@
+package org.launchcode;
+
+import java.time.LocalDate;
+
+public class MenuItem {
+    private String name;
+    private String description;
+    private double price;
+    private String category;
+    private final LocalDate dateAdded;
+
+    public MenuItem(String name, String description, double price, String category) {
+        this.name = name;
+        this.description = description;
+        this.price = price;
+        this.category = category;
+        this.dateAdded = LocalDate.now();
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+
+    //Only getter for dateAdded
+
+    public LocalDate getDateAdded() {
+        return dateAdded;
+    }
+}
